#2012 Extreure les dades del motor a un servei de dades
- Correció d'errors al crear i modificar dades simples al expedient
- Afegida consulta de dades per consultes
- Modificada consulta per generar l'excel de consultes

+++ helium-commons/src/main/java/es/caib/helium/commons/dto/TascaDadaDto.java
@@ -289,7 +289,7 @@ public class TascaDadaDto {
 		if (CampTipusDto.STRING.equals(campTipus)) {
 			return String.class;
 		} else if (CampTipusDto.INTEGER.equals(campTipus)) {
-			return Long.class;
+			return Integer.class;
 		} else if (CampTipusDto.FLOAT.equals(campTipus)) {
 			return Double.class;
 		} else if (CampTipusDto.BOOLEAN.equals(campTipus)) {
